# refactor(sumologicexporter): optimize compressor using sync.Pool