This PR tries to address unnecessary creation of data compressor instances on every `push(Logs|Metrics|Traces)` invocation using `sync.Pool`

```
$ go test -v -count 1 -run ^$ -bench Benchmark_ExporterPushLogs -benchmem -count 5 -benchtime 200x . > old.txt
$ git(5651893) gco -
Previous HEAD position was 5651893 tests(sumologicexporter): add exporter logs sending benchmark
Switched to branch 'optimize-compression'
$ go test -v -count 1 -run ^$ -bench Benchmark_ExporterPushLogs -benchmem -count 5 -benchtime 200x . > new.txt
$ benchcmp old.txt new.txt
benchcmp is deprecated in favor of benchstat: https://pkg.go.dev/golang.org/x/perf/cmd/benchstat
benchmark                         old ns/op     new ns/op     delta
Benchmark_ExporterPushLogs-16     2396758       2079249       -13.25%
Benchmark_ExporterPushLogs-16     2558661       2250972       -12.03%
Benchmark_ExporterPushLogs-16     3170924       2782736       -12.24%
Benchmark_ExporterPushLogs-16     2857462       2505237       -12.33%
Benchmark_ExporterPushLogs-16     3131852       2938191       -6.18%

benchmark                         old allocs     new allocs     delta
Benchmark_ExporterPushLogs-16     24928          24744          -0.74%
Benchmark_ExporterPushLogs-16     24923          24732          -0.77%
Benchmark_ExporterPushLogs-16     24928          24735          -0.77%
Benchmark_ExporterPushLogs-16     24935          24741          -0.78%
Benchmark_ExporterPushLogs-16     24933          24734          -0.80%

benchmark                         old bytes     new bytes     delta
Benchmark_ExporterPushLogs-16     13578172      3824937       -71.83%
Benchmark_ExporterPushLogs-16     13575409      3562536       -73.76%
Benchmark_ExporterPushLogs-16     13575177      3666318       -72.99%
Benchmark_ExporterPushLogs-16     13579142      3653744       -73.09%
Benchmark_ExporterPushLogs-16     13576525      3586314       -73.58%
```